Support unicode operators
The choice of … to mean .. instead of ... might seem odd, but .. is used
a lot more often than ..., and many languages use an ellipsis for ranges
too, so I think it makes more sense this way.

The Letter rule gets a bit uglier because apparently, fragment tokens
match before regular ones regardless of their position in the source
file, so the operators were recognized as letters and attempted to be
parsed as identifiers. Fixed by excluding some unicode blocks and
characters (by splitting the range up).